requirements.txt
anthropic>=0.40.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
pytz>=2024.1
markdown>=3.8
python-dotenv>=1.0.0
